+++ src/mlpack/core/kernels/cosine_distance.hpp
@@ -35,16 +35,6 @@ class CosineDistance
   template<typename VecTypeA, typename VecTypeB>
   static double Evaluate(const VecTypeA& a, const VecTypeB& b);
 
-  /**
-   * Returns a string representation of this object.
-   */
-  std::string ToString() const
-  {
-    std::ostringstream convert;
-    convert << "CosineDistance [" << this << "]" << std::endl;
-    return convert.str();
-  }
-
   //! Serialize the class (there's nothing to save).
   template<typename Archive>
   void Serialize(Archive& /* ar */, const unsigned int /* version */) const { }