+++ src/mlpack/core/kernels/laplacian_kernel.hpp
@@ -98,15 +98,6 @@ class LaplacianKernel
     ar & data::CreateNVP(bandwidth, "bandwidth");
   }
 
-  //! Return a string representation of the kernel.
-  std::string ToString() const
-  {
-    std::ostringstream convert;
-    convert << "LaplacianKernel [" << this << "]" << std::endl;
-    convert << "  Bandwidth: " << bandwidth << std::endl;
-    return convert.str();
-  }
-
  private:
   //! Kernel bandwidth.
   double bandwidth;