+++ src/mlpack/core/math/range.hpp
@@ -161,12 +161,6 @@ class Range
    */
   template<typename Archive>
   void Serialize(Archive& ar, const unsigned int version);
-
-  /**
-   * Returns a string representation of an object.
-   */
-  inline std::string ToString() const;
-
 };
 
 } // namespace math