+++ src/mlpack/core/metrics/ip_metric_impl.hpp
@@ -68,17 +68,6 @@ void IPMetric<KernelType>::Serialize(Archive& ar,
   ar & data::CreateNVP(kernel, "kernel");
 }
 
-// Convert object to string.
-template<typename KernelType>
-std::string IPMetric<KernelType>::ToString() const
-{
-  std::ostringstream convert;
-  convert << "IPMetric [" << this << "]" << std::endl;
-  convert << "  Kernel: " << std::endl;
-  convert << util::Indent(kernel->ToString(), 2);
-  return convert.str();
-}
-
 // A specialization for the linear kernel, which actually just turns out to be
 // the Euclidean distance.
 template<>