+++ src/mlpack/methods/gmm/gmm_impl.hpp
@@ -395,35 +395,6 @@ double GMM<FittingType>::LogLikelihood(
 }
 
 /**
-* Returns a string representation of this object.
-*/
-template<typename FittingType>
-std::string GMM<FittingType>::ToString() const
-{
-  std::ostringstream convert;
-  std::ostringstream data;
-  convert << "GMM [" << this << "]" << std::endl;
-  convert << "  Gaussians: " << gaussians << std::endl;
-  convert << "  Dimensionality: "<<dimensionality;
-  convert << std::endl;
-  // Secondary ostringstream so things can be indented properly.
-  for (size_t ind=0; ind < gaussians; ind++)
-  {
-    data << "Means of Gaussian " << ind << ": " << std::endl
-        << dists[ind].Mean();
-    data << std::endl;
-    data << "Covariances of Gaussian " << ind << ": " << std::endl ;
-    data << dists[ind].Covariance() << std::endl;
-    data << "Weight of Gaussian " << ind << ": " << std::endl ;
-    data << weights[ind] << std::endl;
-  }
-
-  convert << util::Indent(data.str());
-
-  return convert.str();
-}
-
-/**
  * Serialize the object.
  */
 template<typename FittingType>