+++ src/mlpack/methods/lars/lars.hpp
@@ -169,9 +169,6 @@ class LARS
   template<typename Archive>
   void Serialize(Archive& ar, const unsigned int /* version */);
 
-  // Returns a string representation of this object.
-  std::string ToString() const;
-
  private:
   //! Gram matrix.
   arma::mat matGramInternal;