+++ src/mlpack/methods/linear_regression/linear_regression.cpp
@@ -153,11 +153,3 @@ double LinearRegression::ComputeError(const arma::mat& predictors,
 
   return cost;
 }
-
-std::string LinearRegression::ToString() const
-{
-  std::ostringstream convert;
-  convert << "Linear Regression [" << this << "]" << std::endl;
-  convert << "  Lambda: " << lambda << std::endl;
-  return convert.str();
-}